+++ tests/data/expected/chap1/code/under/umixed.py
@@ -2,7 +2,7 @@
 
 a = "In all chapters"   # inline comment
 b = "In chapters 1-3"   # comment on conditional
-c = "In chapters 1-2"   
+c = "In chapters 1-2"
 
 
 for x in range(10):