+++ tests/data/expected/chap5.0/code/after4/amixed.py
@@ -1,7 +1,7 @@
 # This is a sample file
 
 a = "In all chapters"   # inline comment
-d = "In chapters 2 on"  
+d = "In chapters 2 on"
 
 
 for x in range(10):