+++ tests/data/expected/chap5.0/code/maxi.py
@@ -1,1 +1,1 @@
-print("only in chapter 5")    
+print("only in chapter 5")