+++ tests/data/expected/chapFour/code/between24/bmixed.py
@@ -1,7 +1,7 @@
 # This is a sample file
 
 a = "In all chapters"   # inline comment
-d = "In chapters 2 on"  
+d = "In chapters 2 on"
 
 e = "In chapters 3 to 4"  # inline comment
 f = "  as a block"